Move description overrides to be part of edition directly
It's not something we want to encourage other classes to do!
