+++ app/models/edition.rb
@@ -1,6 +1,5 @@
 class Edition < ApplicationRecord
   include SymbolizeJSON
-  include DescriptionOverrides
 
   enum content_store: {
     draft: 'draft',
@@ -217,6 +216,31 @@ class Edition < ApplicationRecord
   def web_url
     return unless base_path
     Plek.current.website_root + base_path
+  end
+
+  # FIXME These are required for the special 'description' column but should
+  # be removed when the column is fixed.
+  def description=(value)
+    super("value" => value)
+  end
+
+  def description
+    super.fetch("value")
+  end
+
+  def attributes
+    attributes = super
+    description = attributes.delete("description")
+
+    if description
+      attributes.merge("description" => description.fetch("value"))
+    else
+      attributes
+    end
+  end
+
+  def self.column_defaults
+    super.merge("description" => nil)
   end
 
 private
