+++ spec/models/edition_spec.rb
@@ -313,7 +313,7 @@ RSpec.describe Edition do
   end
 
   it_behaves_like WellFormedContentTypesValidator
-  it_behaves_like DescriptionOverrides
+  it_behaves_like "description overrides"
 
   context "#publish" do
     it "changes the content_store to live" do
